modified code to actually stop when maintenance mode is on

+++ classes/Api.php
@@ -316,6 +316,7 @@ class Api
 
 		$maxTime = 60 * 1000;
 		while ($timer->stop() < $maxTime) {
+			if (Util::isMaintenanceMode()) return;
 			Db::execute("delete from zz_api_characters where isDirector = ''");
 
 			$allChars = Db::query("select apiRowID, cachedUntil from zz_api_characters where errorCount < 10 and cachedUntil < date_sub(now(), interval 30 second) order by cachedUntil, keyID, characterID limit 1000", array(), 0);
@@ -326,6 +327,7 @@ class Api
 
 			if ($total == 0) sleep(1);
 			else foreach ($allChars as $char) {
+				if (Util::isMaintenanceMode()) return;
 				if ($timer->stop() > $maxTime) return;
 
 				$apiRowID = $char["apiRowID"];
